+++ common/src/main/java/me/rufia/fightorflight/CobblemonFightOrFlight.java
@@ -3,7 +3,6 @@ package me.rufia.fightorflight;
 import com.cobblemon.mod.common.entity.pokemon.PokemonEntity;
 import com.cobblemon.mod.common.pokemon.Pokemon;
 import dev.architectury.registry.ReloadListenerRegistry;
-import me.rufia.fightorflight.compat.CobblemonSizeVariationCompat;
 import me.rufia.fightorflight.config.FightOrFlightCommonConfigModel;
 import me.rufia.fightorflight.config.FightOrFlightMoveConfigModel;
 import me.rufia.fightorflight.config.FightOrFlightVisualEffectConfigModel;
@@ -37,7 +36,7 @@ public class CobblemonFightOrFlight {
     private static FightOrFlightCommonConfigModel commonConfig;
     private static FightOrFlightMoveConfigModel moveConfig;
     private static FightOrFlightVisualEffectConfigModel visualEffectConfig;
-    public static CobblemonSizeVariationCompat sizeVariationCompat;
+    //public static CobblemonSizeVariationCompat sizeVariationCompat;//TODO It should be deprecated, don't forget to remove it.
     //private static TriConsumer<PokemonEntity, Integer, Goal> goalAdder;
 
 
@@ -54,8 +53,8 @@ public class CobblemonFightOrFlight {
     }
 
     public static void init(Predicate<String> modCompatPredicate) {
-        sizeVariationCompat = new CobblemonSizeVariationCompat();
-        sizeVariationCompat.tryLoad(modCompatPredicate);
+        //sizeVariationCompat = new CobblemonSizeVariationCompat();
+        //sizeVariationCompat.tryLoad(modCompatPredicate);
         AutoConfig.register(FightOrFlightCommonConfigModel.class, JanksonConfigSerializer::new);
         AutoConfig.register(FightOrFlightMoveConfigModel.class, JanksonConfigSerializer::new);
         AutoConfig.register(FightOrFlightVisualEffectConfigModel.class, JanksonConfigSerializer::new);