+++ common/src/main/java/me/rufia/fightorflight/CobblemonFightOrFlightClient.java
@@ -1,0 +1,10 @@
+package me.rufia.fightorflight;
+
+import me.rufia.fightorflight.utils.msg.FOFMsg;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class CobblemonFightOrFlightClient {
+    private static List<FOFMsg> CLIENT_MESSAGE_LIST = new ArrayList<>();
+}