+++ common/src/main/java/me/rufia/fightorflight/utils/msg/FOFMsg.java
@@ -1,0 +1,4 @@
+package me.rufia.fightorflight.utils.msg;
+
+public class FOFMsg {
+}